+++ digital_game_nomad/app/board/presenter/BoardPresenter.tsx
